update: cleanups
