+++ src/py/mat3ra/made/tools/analyze/__init__.py
@@ -2,11 +2,12 @@ from typing import Union
 
 import numpy as np
 from mat3ra.made.material import Material
-from mat3ra.made.tools.analyze.other import get_chemical_formula_empirical
-from mat3ra.made.tools.analyze.utils import decorator_perform_operation_in_cartesian_coordinates
-from mat3ra.made.tools.build import MaterialWithBuildMetadata
 from pydantic import BaseModel
 from scipy.spatial.distance import pdist
+
+from .other import get_chemical_formula_empirical
+from .utils import decorator_perform_operation_in_cartesian_coordinates
+from ..build import MaterialWithBuildMetadata
 
 
 class BaseMaterialAnalyzer(BaseModel):
